+++ CHANGELOG.md
@@ -3,3 +3,12 @@
 ## 0.1.0.0 -- YYYY-mm-dd
 
 * First version. Released on an unsuspecting world.
+
+# TODO
+- [ ] Be able to have null if you dont need 3 class periods
+- [ ] Add the actual logic
+    - [x] Check that Classes (not marked with skip_class) and Exams dont overlap
+    - [x] Have a min_rest variable for minimum rest between classes
+        - [x] Same but for exams
+    - [ ] Minimize downtime between classes
+    - [x] Priority indicator for when there is a max allowed classes